Remove original agenda lab
